revert reording
